Move Guideline into own file
