+++ src/glyph/mod.rs
@@ -13,8 +13,9 @@ use std::sync::Arc;
 use druid::{Data, Lens};
 
 use crate::error::{Error, ErrorKind, GlifError, GlifErrorInternal};
+use crate::guideline::{Guideline, Line};
 use crate::names::NameList;
-use crate::shared_types::{Color, Guideline, Identifier, Line, Plist, PUBLIC_OBJECT_LIBS_KEY};
+use crate::shared_types::{Color, Identifier, Plist, PUBLIC_OBJECT_LIBS_KEY};
 
 /// The name of a glyph.
 pub type GlyphName = Arc<str>;
